assert that triangular_window matches e3x
